// src/app/iniciosesion/interface/persona.interface.ts
export interface Persona {
  name: string;
  lastname: string;
  document: string;
  age: number;
  alture: number;
  size: string;
  weight: string;
}

export interface PersonaReponse {
  id: number;
  name: string;
  lastname: string;
  document: string;
  age: number;
  alture: number;
  size: string;
  weight: string;
}